docs: remove 11ty plugin altogether
